requirements.txt
cryptography==41.0.4
flet==0.28.3
fletxr==0.1.4.dev1